requirements.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1